refactor(settings): separate StatisticsSectionView into Screen + pure components
- Extract StatisticsSectionScreen as coordinator (observes ViewModels)
- Make StatisticsSectionView a pure presentation component (Binding only)
- Extract DisabledStatsView as standalone component
- Split PublisherStatsSection and SubscriberStatsSection into own files
- Move shared row components (StatsRow, StatsSectionHeader, StatsSubHeader) to UI/Components/
- Move String.nilIfEmpty extension to Utils/String+NilIfEmpty.swift
