+++ VERA/VERASettings/VERASettings/Utils/String+NilIfEmpty.swift
@@ -1,0 +1,9 @@
+//
+//  Created by Vonage on 22/2/26.
+//
+
+import Foundation
+
+extension String {
+    var nilIfEmpty: String? { isEmpty ? nil : self }
+}
